renaming to distinguish between creating a custom bot and using the custom bot page
